Add files via upload
The program asks for a ciphertext which when entered is separated by commas. It removes the commas and saves the ciphertext as a string in a vector. 
I have made a vector of maps. Or you can call it an array of maps. The idea is that the vector has 4 indexes for every plaintext and each of the indexes has a map that would contain the unit frequency letter and the no of times they appear. but i guess having a vector of maps is not really a good idea. Plus it gives some indexing error.